HardwareTestAutomation/requirements.txt
labjack_ljm==1.23.0
